Header
Building of a responsive navigation.

When the viewport is less than 768 pixels wide a smaller logo and burger menu will be displayed.

If the viewport exceeds 768 pixels, a different logo and quick links will be displayed.

Mobile menu opens with JavaScript and body overflow will be hidden. Content will be implemented later on.