Product information
Building and laying out the page with product specific information.

The collection section will echo the collection linked to the product. On bigger screens a product image will be visible aswell.

Information section displays the modelname and referential number inside the container header.

The container body contains a flickity slider that shows pictures and specific information about the model and calibre it uses.

Any values and media placed on the page is test data. Later on these will be fetched from the database.

A section with information about the collection itself will be made later on. This content will dynamically be loaded into the page depending on the product.