Account registration
Potential customers are able to make an account.

After filling in the sign-up form, submit the registration.

A check will run through all accounts to see if the mail address already exists. If it does exist, you'll be redirected to the form displaying an error.

Is the mail address available? You'll be redirected to the form displaying a success message. The account will be inserted into the account table.