Authentication
Implementing ability to sign in with a registered account.

On submitting the 'sign in' form the mail address will be checked for existence in the database.

When the stored password (hashed) equals the submitted form value, queried account information will be used to start several sessions.

On success the user will be redirected to their profile page (profile.php is made).

When failed, they'll be sent back to account.php. Incorrect password error will be displayed.