Updating profile, Error/Success messages
Error and Success messages have been updated to follow the same base styling on every page, for every message. Placement changed aswell.

Implemented the functionality to update profile information when logged in. After completing the form a success message will be displayed.

When opening the update form the users current values are displayed as placeholders (values).