Logging out & header changes.
When any user is logged in they'll be able to sign out.

Header has gotten some small changes to display the bag, and logout button only when logged in.
The profile button either redirects to account.php (when no session is active) or to profile.php when someone is logged in.