Changing password
Users are now able to change their passwords. The password has to be repeated making sure the user filled in the correct password. If the passwords are not equal, an error message will occur. Are they equal? The user will be redirected to the account.php page to login with their new password.