Edit products and carousel changes.
Implementation of editing the database values for a specific product. Employees will be able to click the edit pencil on any products' page to update it's values through a form. The form will be displayed instead of the product carousel.

The carousel (Flickity) to display product information caused syntax errors. The JavaScript data used inside the HTML tags intervened with PHP code.

Solution was to make a JavaScript file that holds the needed values to make the carousel usable.