Filtering and displaying products.
Implementation of live filtering through checkboxes with jQuery & AJAX. Example taken online and rebuild to be fit for a PHP function, easier to read and easily extendable with more filter values.

Filter values fetched through a function made earlier have changed. Filters fetched with a join now display their ID instead of name in the value attribute. The name will still be echo'd for the enduser.