Deleting a watch.
Products can now be removed. The user will be asked to check the referential number before deletion.

First a query will be run to check whether the product can be found in anyones cart. If so, the product will be deleted from all carts, and then from the product table. If not, it will only be deleted from the product table.

Before echo'ing the success or fail message another check makes sure the product is actually removed from the database or not.